+++ gulpfile.js
@@ -2,6 +2,8 @@
 const gulp = require('gulp');
 const tools = require('./src');
 
-tools.initialize(gulp);
+tools.initialize(gulp, {
+  webpack: false
+});
 
 gulp.task('default', ['pre-release']);